// iina-plugin-thunder-subtitle-ui/src/define/result.d.ts
// Generated by https://quicktype.io

export interface SubtitleItem {
  source: number
  score: number
  languages: string[]
  cid: string
  fingerprintf_score: number
  ext: EXT
  extra_name: ExtraName
  duration: number
  gcid: string
  name: string
  url: string
}

export enum EXT {
  Ass = 'ass',
  Srt = 'srt',
}

export enum ExtraName {
  网友上传 = '（网友上传）',
}